+++ internal/proxy/bin/bin_assets.go
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b5652cb40c85f1be7f09ef651e857eed3d2e62d3fec74318c078b3288dc56ec7
-size 14489897