+++ platform/lambda/lambda.go
@@ -293,13 +293,13 @@ func (p *Platform) ApplyStack(region string) error {
 }
 
 // populateZones fetches the existing hosted zones, storing
-// the HostedZoneId if present. This is required because domains
-// purchased via Route53 will already have a hosted zone,
-// so we need to reference it instead of creating a new zone.
+// the HostedZoneId if present.
 //
-// Currently subdomains are stripped. For example if you create
-// an app which maps only production to "api.example.com", the
-// zone "example.com" will be created (or discovered).
+// This is required because domains purchased via Route53
+// will already have a hosted zone, and we may already have
+// a hosted zone for the apex (example.com) and we are
+// creating a sub-domain (api.example.com), these should
+// all live in the same zone.
 func (p *Platform) populateZones() error {
 	r := route53.New(session.New(aws.NewConfig()))
 
@@ -312,12 +312,11 @@ func (p *Platform) populateZones() error {
 	}
 
 	for _, s := range p.config.Stages.List() {
-		domain := util.Domain(s.Domain)
-		log.Debugf("finding stage dns zones for %s %s (%s)", s.Name, domain, s.Domain)
+		log.Debugf("finding stage dns zones for %s %s", s.Name, s.Domain)
 		for _, z := range res.HostedZones {
-			if domain+"." == *z.Name {
+			if strings.Contains(s.Domain+".", *z.Name) {
 				s.HostedZoneID = strings.Replace(*z.Id, "/hostedzone/", "", 1)
-				log.Debugf("found existing dns zone %s (%s) mapped to stage %s", s.Domain, s.HostedZoneID, s.Name)
+				log.Debugf("found existing dns zone %s (%s)", *z.Name, s.HostedZoneID)
 			}
 		}
 	}
@@ -352,18 +351,9 @@ func (p *Platform) createCerts() error {
 			continue
 		}
 
-		domain := util.Domain(s.Domain)
-
-		// already requested
-		if util.StringsContains(domains, domain) {
-			log.Debugf("already requested cert for %s", domain)
-			continue
-		}
-
-		// see if the cert exists, stripping any subdomain
-		// since we request only a single wildcard cert.
+		// see if the cert exists
 		log.Debugf("looking up cert for %s", s.Domain)
-		arn := getCert(res.CertificateSummaryList, domain)
+		arn := getCert(res.CertificateSummaryList, s.Domain)
 		if arn != "" {
 			log.Debugf("found cert for %s: %s", s.Domain, arn)
 			s.Cert = arn
@@ -372,15 +362,14 @@ func (p *Platform) createCerts() error {
 
 		// request the cert
 		res, err := a.RequestCertificate(&acm.RequestCertificateInput{
-			DomainName:              &domain,
-			SubjectAlternativeNames: aws.StringSlice([]string{"*." + domain}),
+			DomainName: &s.Domain,
 		})
 
 		if err != nil {
 			return errors.Wrapf(err, "requesting cert for %s", s.Domain)
 		}
 
-		domains = append(domains, domain)
+		domains = append(domains, s.Domain)
 		s.Cert = *res.CertificateArn
 	}